Refactoring testů
- vytvořena TestCase, která vytváří kontejner
- úprava strktury složek
