+++ .gitignore
@@ -2,4 +2,4 @@
 /vendor
 /tests/**/*.actual
 /tests/**/*.expected
-
+/tests/temp
