+++ composer.json
@@ -30,5 +30,10 @@
 		"psr-4": {
 			"Pd\\Version\\": "src/"
 		}
+	},
+	"autoload-dev": {
+		"classmap": [
+			"tests/"
+		]
 	}
 }
