--- tests/DI/cache/.gitignore
@@ -1,3 +1,0 @@
-*
-
-!.gitignore
