+++ tests/PdTests/Version/DI/extension.neon
